# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "School Management System",
    "summary": "Multi-tenant school management platform",
    "description": "EduCloud API -- identity, tenancy, RBAC, plans and invitations.\n\n**Tenant isolation** is enforced by PostgreSQL Row-Level Security. The `org` claim in your access token scopes every query at the database level; cross-tenant reads return 404, never 403, so resource existence is never disclosed.\n\n**School scoping** is a second, softer boundary inside the organization: a school-scoped membership sees only its own campus, while an organization owner spans all of them.\n\n**Permissions** are `resource:action` strings resolved per request from the caller's role. They are deliberately not embedded in the token, so revoking access takes effect on the very next request.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/org/usage": {
      "get": {
        "tags": [
          "Organization"
        ],
        "summary": "Get Usage",
        "description": "Live counters against plan limits (spec \u00a78).\n\nReadable by ANY authenticated member, deliberately NOT gated behind\n`billing:read`. A teacher who hits a student limit needs to understand why the\ncreate failed; answering only \"forbidden\" sends them to ask the principal a\nquestion the UI could have answered itself.",
        "operationId": "get_usage_api_v1_org_usage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/billing/subscription": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "Get Subscription",
        "operationId": "get_subscription_api_v1_billing_subscription_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionRead"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/billing/subscribe": {
      "post": {
        "tags": [
          "Billing"
        ],
        "summary": "Subscribe",
        "description": "Move onto a paid plan.\n\nShares an implementation with `change-plan`: from the system's point of view\n\"subscribe\" and \"change plan\" are the same transition, since every organization\nalready holds a free-plan subscription row. Kept as two endpoints because they\nare two different intentions in the UI, and the audit trail reads better for it.",
        "operationId": "subscribe_api_v1_billing_subscribe_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionRead"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/billing/change-plan": {
      "post": {
        "tags": [
          "Billing"
        ],
        "summary": "Change Plan",
        "description": "Upgrade or downgrade. A downgrade below current usage never deletes data.",
        "operationId": "change_plan_api_v1_billing_change_plan_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionRead"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/billing/cancel": {
      "post": {
        "tags": [
          "Billing"
        ],
        "summary": "Cancel Subscription",
        "description": "Cancel at period end. Immediate cancellation is a super-admin action.",
        "operationId": "cancel_subscription_api_v1_billing_cancel_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionRead"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/billing/invoices": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "List Invoices",
        "description": "This organization's invoices.\n\nNo explicit `WHERE organization_id` filter: RLS applies it at the database, and\nadding a redundant one here would suggest the filter is what provides the\nisolation. It is not -- the policy is.",
        "operationId": "list_invoices_api_v1_billing_invoices_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvoiceRead"
                  },
                  "type": "array",
                  "title": "Response List Invoices Api V1 Billing Invoices Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
    "/api/v1/billing/invoices/{invoice_id}/pdf": {
      "get": {
        "tags": [
          "Billing"
        ],
        "summary": "Download Invoice Pdf",
        "description": "Generate one invoice after RLS and permission checks have resolved it.",
        "operationId": "download_invoice_pdf_api_v1_billing_invoices__invoice_id__pdf_get",
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Invoice Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "422": {
            "description": "Validation Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BillingCycle": {
        "type": "string",
        "enum": [
          "monthly",
          "yearly"
        ],
        "title": "BillingCycle"
      },
      "CancelRequest": {
        "properties": {
          "at_period_end": {
            "type": "boolean",
            "title": "At Period End",
            "default": true
          }
        },
        "type": "object",
        "title": "CancelRequest"
      },
...
      "InvoiceRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "number": {
            "type": "string",
            "title": "Number"
          },
          "amount_subtotal": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount Subtotal"
          },
          "amount_tax": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount Tax"
          },
          "amount_total": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount Total"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "issued_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issued At"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "paid_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "number",
          "amount_subtotal",
          "amount_tax",
          "amount_total",
          "currency",
          "status",
          "issued_at",
          "due_at",
          "paid_at",
          "pdf_url",
          "created_at"
        ],
        "title": "InvoiceRead"
      },
...
      "SubscribeRequest": {
        "properties": {
          "plan_code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Plan Code"
          },
          "billing_cycle": {
            "$ref": "#/components/schemas/BillingCycle",
            "default": "monthly"
          }
        },
        "type": "object",
        "required": [
          "plan_code"
        ],
        "title": "SubscribeRequest"
      },
      "SubscriptionRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "plan_code": {
            "type": "string",
            "title": "Plan Code"
          },
          "plan_name": {
            "type": "string",
            "title": "Plan Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "billing_cycle": {
            "type": "string",
            "title": "Billing Cycle"
          },
          "trial_ends_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial Ends At"
          },
          "current_period_start": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period Start"
          },
          "current_period_end": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Period End"
          },
          "cancel_at_period_end": {
            "type": "boolean",
            "title": "Cancel At Period End"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "plan_code",
          "plan_name",
          "status",
          "billing_cycle",
          "trial_ends_at",
          "current_period_start",
          "current_period_end",
          "cancel_at_period_end",
          "cancelled_at"
        ],
        "title": "SubscriptionRead"
      },
...
      "UsageItem": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "current": {
            "type": "integer",
            "title": "Current"
          },
          "allowed": {
            "type": "integer",
            "title": "Allowed"
          },
          "remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining"
          },
          "is_unlimited": {
            "type": "boolean",
            "title": "Is Unlimited"
          },
          "is_exhausted": {
            "type": "boolean",
            "title": "Is Exhausted"
          }
        },
        "type": "object",
        "required": [
          "key",
          "current",
          "allowed",
          "remaining",
          "is_unlimited",
          "is_exhausted"
        ],
        "title": "UsageItem",
        "description": "One metered limit and its consumption. `allowed = -1` means unlimited."
      },
      "UsageResponse": {
        "properties": {
          "plan_code": {
            "type": "string",
            "title": "Plan Code"
          },
          "organization_status": {
            "type": "string",
            "title": "Organization Status"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/UsageItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "plan_code",
          "organization_status",
          "items"
        ],
        "title": "UsageResponse",
        "description": "`GET /org/usage` -- live counters against plan limits."
      },
...
    }
  }
}